requirements.txt
Flask_Login==0.4.1
Flask==0.12.2
Werkzeug==0.14.1
WTForms==2.1
SQLAlchemy==1.2.8
Flask_Script==2.0.6
Flask_WTF==0.14.2
Flask_SQLAlchemy==2.3.2
alembic==0.9.10
flask_bootstrap==3.3.7.1
flask_migrate==2.2.1
flask_moment==0.6.0
flask_pagedown==0.2.2
flask_paginate==0.5.1
Flask_Bcrypt==0.7.1
Flask-Markdown
pymysql